+++ sr_port/timeout2msec.c
@@ -3,6 +3,9 @@
  * Copyright (c) 2001-2017 Fidelity National Information	*
  * Services, Inc. and/or its subsidiaries. All rights reserved.	*
  *								*
+ * Copyright (c) 2020 YottaDB LLC and/or its subsidiaries.	*
+ * All rights reserved.						*
+ *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
  *	under a license.  If you do not know the terms of	*
@@ -29,7 +32,7 @@
  *			else		timeout * 1000 (timeout value in milliseconds)
  */
 
-int4	timeout2msec (int4 timeout)
+int4	timeout2msec (uint8 timeout)
 {
 
 	/* less than 0, make it 0; if multiplying by 1000 causes arithmetic overflow, cap the value */
